OutputSaver working with create_catalog_entry
